pyproject.toml
[project]
name = "adk-a2a-gke-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk[a2a]>=1.21.0",
]
